SCLang: Check nan is not a tag in slot
There are many safe nan values, but a few are not safe.
The server's Convolution3 is a known source of unsafe nans.
Supercollider VM itself does not generate unsafe nans.
Therefore, in future, it might be possible to remove this check from here and place it at external boundaries instead.